Re-architect login process (#15)
This commit re-architects the login process to:

1. Make sure that the live socket is not open at all unless the user
   is correctly logged in.
2. Have the login link on a separate route, with a dedicated page that
   can be used to explain permission requirements, etc.
3. Remove all authentication checks from explorer live templates, as
   they can now assume that the user is always authenticated.

LiveView doesn't offer a built-in way to pass data from the connection
assigns to the live socket assigns, so the 'mount/3' callback cannot be
streamlined at this stage and has to repeat the authentication operation
to get access to session id and user. This is why mock counts in test
have increased by 1.