Extract mini player/progress bar to their own components
This commit greatly reduces the amount of data sent over the wire due to
changes in the current player.

Before this commit, the entire explorer view would re-render anytime the
currently playing item progress was updated (roughly every second).

This caused a lot of bandwidth waste, as the only component concerned
with progress updates is the progress bar included in the miniplayer.

With this commit, the main ui is re-rendered only when either the
playing item changes or playback is played/paused/stopped.

Changes include:

- The progress bar template fragment is promoted to a LiveComponent.
- The ProgressBar component is rendered with an ID, which makes it
stateful. As it's stateful, it's not removed and re-rendered when the
parent changes.
- Changes in the ProgressBar component are triggered by explicitly
sending an update via 'send_update/2` from its parent ExplorerLive view.
- When receiving now playing updates, the ExplorerLive component
distinguishes between progress updates (which are sent to ProgressBar)
and other updates, which re-render the rest of the UI.

To extract markup containing the progress bar, it’s necessary to create
a miniplayer component as it’s not possible to call a live component
from a normal template.