Restructure header to re-enable live search and home
This commit promotes the main header to its own view/templates, so that
its individual elements can be rendered where appropriate.

For logged out viewers, we can inline the logo and main header markup
structure in the app template.

For logged in viewers, header components are rendered directly by the
explorer live template.

With this structure we can:

- Re-enable live-powered search (quicker)
- Implement logo -> home navigation as live (quicker)
- Remove the authentication concern from the layout view - non-live maps
  to logged-out, live maps to logged-in