Enable browser player (#48)
This commit adds an invisible audio player and a volume control slider.

The audio player uses a minimal subset of the Spotify Web Player SDK
functionality. It reuses the session token already used for all other API
calls and it's kept up to date via a subscription that broadcasts the new token
when the old one expires.

At this point in time, it's only possible to have one Web Player open at any
given time. Still, the player is assigned a pronounceable, random name for
easier identification.

The volume slider is a simple range input, with some accessibility attributes
and custom styling.